refactor: share attachment serialization helpers between google-drive and onedrive plugins
Move serializeDocAttachments, deserializeDocAttachments and the
inline-stripForComparison logic (now stripAllAttachmentDataForComparison)
from each plugin's document-handling.ts / upstream.ts into
src/replication-protocol/helper.ts so both replication backends share
a single implementation. The plugins re-export the same names so the
public API is unchanged.
